requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.13.0
jupyter>=1.0.0
nba_api>=1.11.0
joblib>=1.3.0
tabulate>=0.9.0
requests-cache>=1.1.0
